==> r_scripts/0_merge.Rmd <==
---
title: "Triple difference merge datasets"
author: "Andreas Olden"
date: "Updated: `r format(Sys.time(), '%m/%d/%Y')`"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, warning = FALSE)
```

```{r}
#install.packages("renv")
#renv::init()
renv::restore()
```


```{r load packages, include = FALSE}
#list.of.packages <- c("tidyverse", "here", "haven", "foreach", "doParallel", "doRNG", "fixest", "kableExtra", "renv")
#new.packages <- list.of.packages[!(list.of.packages %in% installed.packages()[,"Package"])]
#if(length(new.packages)) install.packages(new.packages)

library(tidyverse) # needs no introduction
library(here) #relative file paths
library(haven)
#library(kableExtra)

```

# Intro

This is the file that generates the data material for the simulation. Raw material downloaded: 19 November 2020 from: https://www.nber.org/research/data/current-population-survey-cps-data-nber

## We extract: 

* minsamp   chr "Month in sample (4 & 8 are departing)"
* intmonth   chr "Interview month"
* hhid   chr "Household ID (12 digits)"
* state   chr "State"
* age     chr "Age" 
* sex      chr "Sex" chr [1:2] "Male" "Female"
* uearnwk chr "(dp) Usual earnings per week"
* earnwke  chr "(e&dp) Earnings per week" fra og med 1994
* year    

## Read data

```{r}
df79 <- read_dta( here( "r_scripts/data/raw_data", "morg79.dta") ) %>% 
  select(minsamp, intmonth, hhid, state, age, sex, uearnwk, year)

df80 <- read_dta( here( "r_scripts/data/raw_data", "morg80.dta") ) %>% 
  select(minsamp, intmonth, hhid, state, age, sex, uearnwk, year)

df81 <- read_dta( here( "r_scripts/data/raw_data", "morg81.dta") ) %>% 
  select(minsamp, intmonth, hhid, state, age, sex, uearnwk, year)

df82 <- read_dta( here( "r_scripts/data/raw_data", "morg82.dta") ) %>% 
  select(minsamp, intmonth, hhid, state, age, sex, uearnwk, year)

df83 <- read_dta( here( "r_scripts/data/raw_data", "morg83.dta") ) %>% 
  select(minsamp, intmonth, hhid, state, age, sex, uearnwk, year)

df84 <- read_dta( here( "r_scripts/data/raw_data", "morg84.dta") ) %>% 
  select(minsamp, intmonth, hhid, state, age, sex, uearnwk, year)

df85 <- read_dta( here( "r_scripts/data/raw_data", "morg85.dta") ) %>% 
  select(minsamp, intmonth, hhid, state, age, sex, uearnwk, year)

df86 <- read_dta( here( "r_scripts/data/raw_data", "morg86.dta") ) %>% 
  select(minsamp, intmonth, hhid, state, age, sex, uearnwk, year)

df87 <- read_dta( here( "r_scripts/data/raw_data", "morg87.dta") ) %>% 
  select(minsamp, intmonth, hhid, state, age, sex, uearnwk, year)

df88 <- read_dta( here( "r_scripts/data/raw_data", "morg88.dta") ) %>% 
  select(minsamp, intmonth, hhid, state, age, sex, uearnwk, year)

df89 <- read_dta( here( "r_scripts/data/raw_data", "morg89.dta") ) %>% 
  select(minsamp, intmonth, hhid, state, age, sex, uearnwk, year)

df90 <- read_dta( here( "r_scripts/data/raw_data", "morg90.dta") ) %>% 
  select(minsamp, intmonth, hhid, state, age, sex, uearnwk, year)

df91 <- read_dta( here( "r_scripts/data/raw_data", "morg91.dta") ) %>% 
  select(minsamp, intmonth, hhid, state, age, sex, uearnwk, year)

df92 <- read_dta( here( "r_scripts/data/raw_data", "morg92.dta") ) %>% 
  select(minsamp, intmonth, hhid, state, age, sex, uearnwk, year)

df93 <- read_dta( here( "r_scripts/data/raw_data", "morg93.dta") ) %>% 
  select(minsamp, intmonth, hhid, state, age, sex, uearnwk, year)

################# NEW ENCODING ########################################

df94 <- read_dta( here( "r_scripts/data/raw_data", "morg94.dta") ) %>% 
  rename(uearnwk = earnwke) %>%
  select(minsamp, intmonth, hhid, state, age, sex, uearnwk, year) 

df95 <- read_dta( here( "r_scripts/data/raw_data", "morg95.dta") ) %>% 
  rename(uearnwk = earnwke) %>%
  select(minsamp, intmonth, hhid, state, age, sex, uearnwk, year) 

df96 <- read_dta( here( "r_scripts/data/raw_data", "morg96.dta") ) %>% 
  rename(uearnwk = earnwke) %>%
  select(minsamp, intmonth, hhid, state, age, sex, uearnwk, year) 

df97 <- read_dta( here( "r_scripts/data/raw_data", "morg97.dta") ) %>% 
  rename(uearnwk = earnwke) %>%
  select(minsamp, intmonth, hhid, state, age, sex, uearnwk, year) 

df98 <- read_dta( here( "r_scripts/data/raw_data", "morg98.dta") ) %>% 
  rename(uearnwk = earnwke) %>%
  select(minsamp, intmonth, hhid, state, age, sex, uearnwk, year) 

df99 <- read_dta( here( "r_scripts/data/raw_data", "morg99.dta") ) %>% 
  rename(uearnwk = earnwke) %>%
  select(minsamp, intmonth, hhid, state, age, sex, uearnwk, year) 


df_merged <- rbind(
  df79, df80, df81, df82, df83, df84, df85, df86, df87, df88, df89, df90, df91, df92, df93, df94, df95, df96, df97, df98, df99
  )

rm(
  df79, df80, df81, df82, df83, df84, df85, df86, df87, df88, df89, df90, df91, df92, df93, df94, df95, df96, df97, df98, df99)

```

# Skim resulting DF
```{r}
glimpse(df_merged)
```

# Save df_merged
```{r}
write_rds(df_merged, here("r_scripts/data/cleaned_data", "df_merged.rds"))
```


# state codes

- 63    Alabama                                                        
- 94    Alaska                                                         
- 86    Arizona                                                        
- 71    Arkansas                                                       
- 93    California                                                     
- 84    Colorado                                                       
- 16    Connecticut                                                     
- 51    Delaware                                                       
- 53    District of Columbia                                           
- 59    Florida                                                        
- 58    Georgia                                                         
- 95    Hawaii                                                         
- 82    Idaho                                                          
- 33    Illinois                                                        
- 32    Indiana                                                        
- 42    Iowa                                                           
- 47    Kansas                                                         
- 61    Kentucky                                                        
- 72    Louisiana                                                      
- 11    Maine                                                          
- 52    Maryland                                                       
- 14    Massachusetts                                                  
- 34    Michigan                                                       
- 41    Minnesota                                                      
- 64    Mississippi                                                     
- 43    Missouri                                                       
- 81    Montana                                                        
- 46    Nebraska                                                       
- 88    Nevada                                                          
- 12    New Hampshire                                                  
- 22    New Jersey                                                     
- 85    New Mexico                                                      
- 21    New York                                                       
- 56    North Carolina                                                 
- 44    North Dakota                                                   
- 31    Ohio                                                            
- 73    Oklahoma                                                       
- 92    Oregon                                                         
- 23    Pennsylvania                                                   
- 15    Rhode Island                                                   
- 57    South Carolina                                                 
- 45    South Dakota                                                   
- 62    Tennessee                                                       
- 74    Texas                                                          
- 87    Utah                                                           
- 13    Vermont                                                        
- 54    Virginia                                                        
- 91    Washington                                                     
- 55    West Virginia                                                  
- 35    Wisconsin                                                       
- 83    Wyoming
