==> r_scripts/3_results.Rmd <==
---
title: "Results"
author: "Andreas Olden"
date: "Updated: `r format(Sys.time(), '%m/%d/%Y')`"
output: html_document
---

```{r}
#renv::init()
renv::restore()
```

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE)
options(scipen = 999)
```

```{r load packages, include = FALSE}
library(tidyverse)
library(here)
library(kableExtra)
```

```{r define functions and load data, include = FALSE}
lessthan = function(x) {x<0.05}
rejection_rates = function(x) {
  x %>% lapply(., lessthan) %>% lapply(., sum) %>% as.data.frame(.)/10000
}

r1_full = rejection_rates(as.data.frame(read_rds(here("r_scripts/sim_res", "sim_25.rds")) ))
r2_full = rejection_rates(as.data.frame(read_rds(here("r_scripts/sim_res", "sim_5.rds")) ))
r3_full = rejection_rates(as.data.frame(read_rds(here("r_scripts/sim_res", "sim_2.rds")) ))
r4_full = rejection_rates(as.data.frame(read_rds(here("r_scripts/sim_res", "sim_1.rds")) ))
```

```{r read nobs, include = FALSE}
df_nobs = readRDS( here("r_scripts/sim_res", "df_nobs.rds"))
```


# Table 4 (page 22)

```{r tab 1}
rbind("25"=r1_full, "5"=r2_full, 
      "2"=r3_full, "1"=r4_full) %>% 
  select(did_ind_s, did_ind_w, did_ind_c, trip_ind_s, trip_ind_w, trip_ind_c) %>%
  kbl(caption = "No effect: Rejection rates for individual level data models",
      col.names = c( "Standard", "White", "Cluster", "Standard", "White", "Cluster"),
      align = "c"
      ) %>%
  kable_classic(full_width = F) %>%
  add_header_above(
    c(" " = 1, "(1)"= 1, "(2)"= 1, "(3)"= 1, "(4)"= 1, "(5)"= 1, "(6)"= 1)
    ) %>%
  add_header_above(
    c(" " = 1, "Difference-in-differences"= 3, "Triple difference" = 3)
    )

```

```{r tab 1 latex, paged.print = FALSE, include = TRUE}
rbind("25"=r1_full, "5"=r2_full, 
      "2"=r3_full, "1"=r4_full) %>% 
  select(did_ind_s, did_ind_w, did_ind_c, trip_ind_s, trip_ind_w, trip_ind_c) %>% 
  kbl(caption = "No effect: Rejection rates for individual level data models",
      col.names = c( "Standard", "White", "Cluster", "Standard", "White", "Cluster"),
      align = "c",
      format = "latex", 
      booktabs = TRUE
      ) %>%
  add_header_above(
    c(" " = 1, "(1)"= 1, "(2)"= 1, "(3)"= 1, "(4)"= 1, "(5)"= 1, "(6)"= 1)
    ) %>%
  add_header_above(
    c(" " = 1, "Difference-in-differences"= 3, "Triple difference" = 3)
    ) 
```

# Table 5 page 23

```{r table 2}
rbind("25"=r1_full, "5"=r2_full, 
      "2"=r3_full, "1"=r4_full) %>% 
  select(did_ind_c, trip_ind_c, w2_did_ind_c, w2_trip_ind_c, w5_did_ind_c, w5_trip_ind_c) %>%
  kbl(caption = "Rejection rates for individual level data models",
      col.names = c( "DiD", "Triple", "DiD", "Triple", "DiD", "Triple"),
      align = "c"
      ) %>%
  kable_classic(full_width = F, html_font = "Cambria") %>%
  add_header_above(
    c(" " = 1, "(1)"= 1, "(2)"= 1, "(3)"= 1, "(4)"= 1, "(5)"= 1, "(6)"= 1)
    ) %>%
  add_header_above(
    c(" " = 1, "No effect"= 2, "2 percent effect" = 2, "5 percent effect" = 2)
    )


```

```{r tab 2 latex, paged.print = FALSE, include = TRUE}
rbind("25"=r1_full, "5"=r2_full, 
      "2"=r3_full, "1"=r4_full) %>% 
  select(did_ind_c, trip_ind_c, w2_did_ind_c, w2_trip_ind_c, w5_did_ind_c, w5_trip_ind_c) %>%
  kbl(caption = "Rejection rates for individual level data models",
      col.names = c( "DiD", "Triple", "DiD", "Triple", "DiD", "Triple"),
      align = "c",
      format = "latex", 
      booktabs = TRUE
      ) %>%
  add_header_above(
    c(" " = 1, "(1)"= 1, "(2)"= 1, "(3)"= 1, "(4)"= 1, "(5)"= 1, "(6)"= 1)
    ) %>%
  add_header_above(
    c(" " = 1, "No effect"= 2, "2 percent effect" = 2, "5 percent effect" = 2)
    )
```

# Table 6 page 24

```{r table 3}
rbind("25"=r1_full, "5"=r2_full, 
      "2"=r3_full, "1"=r4_full) %>% 
  select(did_agg_c, rel_agg_c, trip_agg_c, w2_did_agg_c, w2_rel_agg_c, w2_trip_agg_c, w5_did_agg_c, w5_rel_agg_c, w5_trip_agg_c) %>%
  kbl(caption = "Rejection rates for aggregated models",
      col.names = c( "DiD", "Triple as DiD", "Triple", "DiD", "Triple as DiD", "Triple",  "DiD", "Triple as DiD", "Triple"),
      align = "c"
      ) %>%
  kable_classic(full_width = F, html_font = "Cambria") %>%
  add_header_above(
    c(" " = 1, "(1)"= 1, "(2)"= 1, "(3)"= 1, "(4)"= 1, "(5)"= 1, "(6)"= 1, "(7)"= 1, "(8)"= 1, "(9)"= 1)
    ) %>%
  add_header_above(
    c(" " = 1, "No effect"= 3, "2 percent effect" = 3, "5 percent effect" = 3)
    ) 
```
```{r tab 3 latex, paged.print = FALSE, include = TRUE}
rbind("25"=r1_full, "5"=r2_full, 
      "2"=r3_full, "1"=r4_full) %>% 
  select(did_agg_c, rel_agg_c, trip_agg_c, w2_did_agg_c, w2_rel_agg_c, w2_trip_agg_c, w5_did_agg_c, w5_rel_agg_c, w5_trip_agg_c) %>%
  kbl(caption = "Rejection rates for aggregated models",
      col.names = c("DiD", "Triple as DiD", "Triple", "DiD", "Triple as DiD", "Triple",  "DiD", "Triple as DiD", "Triple"),
      align = "c",
      format = "latex", 
      booktabs = TRUE
      ) %>%
  kable_classic(full_width = F, html_font = "Cambria") %>%
  add_header_above(
    c(" " = 1, "(1)"= 1, "(2)"= 1, "(3)"= 1, "(4)"= 1, "(5)"= 1, "(6)"= 1, "(7)"= 1, "(8)"= 1, "(9)"= 1)
    ) %>%
  add_header_above(
    c(" " = 1, "No effect"= 3, "2 percent effect" = 3, "5 percent effect" = 3)
    ) 


```

# Number of observations 

Note that the number of observations follow the estimation method as their are transformations of the same underlying dataset.

```{r}

as.data.frame(df_nobs) %>% 
  summarise_all(list(mean)) %>% 
  kbl( ) %>% 
  kable_classic(full_width = F, html_font = "Cambria")
```
